DESTDIR is not the same as PREFIX. DESTDIR: http://www.gnu.org/prep/standards/html_node/DESTDIR.html PREFIX: http://www.gnu.org/prep/standards/html_node/Directory-Variables.html#Directory-Variables
When the system is not installed in '/' DESTDIR should be set to a value other than '/'.
PREFIX defines the directory where the binaries should be run from.

I added a simple uninstall to allow repeated ins
PREFIX defines the directory where the binaries should be run from.

I added a simple uninstall to allow repeated installations.