Do not reformat fence/resource agent's longdesc
Agent's long description contains a long text that is often formatted
using line breaks to make it easier to read.  It is better to keep
the format untouched even if it does not fit the terminal / text box
precisely rather then reformat it to make it fit the terminal and make
it far less readable in the process.