Keep cluster quorate when stopping nodes
Stop pacemaker on the nodes, wait for it to finish on all nodes, then
stop corosync on the nodes. This way the cluster remains quorate through
pacemaker and resources shutdown.
Revert patch putting nodes to standby on cluster stop as it is not
a recommended way to shutdown cluster and it will not fix cluster
stop issues.