Fix corosync reload when adding / removing nodes
Previously the config has been reloaded in the process of updating
the config across a cluster. Now it is reloaded when all nodes have
the new config. Reload is run on a single node as corosync will ask
all nodes to reload the config on its own.

Moreover cluster.conf has not been reloaded at all.