Fix Multiarch support in Debian and derivatives
This commit detects the currently used Debian userland's primary
architecture, instead of the currently used kernel, which could be
user-compiled/hoster-provided or from another distro.

It should(R) now work for any architecture in Debian, and also in Debian
chroots/containers running on other architectures and/or distributions,
even when using qemu-user.