web UI: correct handling of timeout when removing multiple resources
If there are only 2 nodes in cluster (timeout for request is 30 seconds per node)
and removing resources takes longer then 1 minute, we have no other nodes to try
remove resources so it will be returned to javascript as there was no response
from cluster. Now we handle this stituation properly.