Tests: handle change in an error message in lxml 3.7
lxml commit abf54e0d9ad306e17d2dd244782b04a945caf633 (landed in
3.7.0) changes the string representation of XMLSyntaxError,
because it gives it `filename` and `lineno` attributes which
its parent class `SyntaxError` uses for the string. This makes
the tests look for the more detailed message when the lxml
version is 3.7.0.0 or higher.

An alternative here would be to use `err.msg`, not `str(err)`,
when raising `LibraryError` from `XMLSyntaxError`; that should
give the same result for all lxml versions. Let me know if
you'd prefer that.