move adding of missing monitoring op to res. agent
Move the responsibility for completeness of necessary operations into
the resource agent module. It is not an ideal place but currently it is
the better place.

It slightly changes behaviour in the resource create command (and some
others) when --no-default-opt is there. Before the implicitly added
monitor operation was created only with default interval 60.
Now is monitor operation taken from the resource agent if is there.