separate node flags: "offline" vs "failed request"
When an action on a cluster node is unsucessful the flag "--force" is
currently desired. When it is not possible to run an action on a node
because the node is offline the flag "--skip-offline" is currently
desired.

This commit duplicates the flag for skipping/forcing the imposibility of
running an action on a node. So the desired separation is now possible.