turn off configs syncing if the nodelist is empty
This fixes problems on Debian which distributes corosync with a default
corosync.conf containing no nodes. The configs synchronization mechanism
then sends configs to (and load from) an empty set of nodes effectively
not saving the configs anywhere. This is especially bad when running
'pcs cluster auth' before creating a cluster since the tokens are
dropped which prevents a cluster to be created and the default
corosync.conf replaced by our own with a valid nodelist.

https://github.com/ClusterLabs/pcs/issues/153