fix booth name
Booth instance name is passed form CLI to LIB by two means: as a command
parameter and in (booth) environment. To make booth commands' interface
consistent, we need to drop one of the ways. Dropping the booth env is a
preferable solution but it cannot be done right now. So we temporarily
go the other way.