reorganize 'resource group add' validators
pcs.lib.command.resource.group_add was producing report items when
searching for elements based on the input ids. This partially conflicted
with group add validator. There were several tight and fragile
connections between the group add command and validator.

To resolve this, searching for the elements was moved to the validator.
Thus, all the fragile connections disappeared as the validator is now a
class and the connections are encapsulated inside of it.