fix translating 0 in reports
Previously, the report COROSYNC_ADDRESS_IP_VERSION_WRONG_FOR_LINK with info
{'address': '10.37.166.132', 'expected_address_type': 'IPv6', 'link_number': 0}
was translated to "Address '10.37.166.132' cannot be used in the link because
the link uses IPv6 addresses" due to integer 0 being considered an empty value.
The translation was missing the link number even if it was specified. It worked
fine with link numbers other than 0.