tests: adapt constraints tests to pacemaker 2.0.4
In pacemaker 2.0.4, support for rules with some specific date-spec
options has been added to 'crm_rule --check' command. As we need to also
support older versions of pacemaker, and these tests are mainly used for
ensuring pcs is displaying constraints/rules correctly, we decided
to increase years used in pcs test so rules are not marked as expired.