Add integration tests and change one TaskResult parameter name
This commit adds tier0 integration tests. It also implements reusable
mocks in the form of mixins. These mocks are also used throughout other
applicable tier0 tests.

During the use of get_task, I was constantly missing the kill_reason
parameter which was previously called kill_requested which
was not descriptive enough to contain the reason to kill the task.