build: deal better with network bottle necks
simply increase the amount of retries and timeouts when
operating in less ideal network environment, that could
cause gems to fail to download, and reduce the number
of parallel downloads to 1.

issue was spotted in CI where many machines would access
rubygem at the same time from one single IP address
and hit some load-balancer limit.

this should not affect the single user as is, but only
build farms

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>