deprecate unused reports and prevent accidental report code reuse
Redeclaring a new report message with a report code that was already
used in the past could lead to PCS crashing if the older report
appears in newer PCS version where it was redeclared. It is no longer
possible to pass mypy check if the deprecated report code is reused.

Along with these mechanism, first unused or replaced reports were
deprecated. These reports can still be displayed by PCS, but their
message can no longer be modified.