add check for closed stdin
Authors of scripts sometimes close stdin. In some commands, we ask the
user for input but first we check that pcs is not invoked from a
script. This check was returning a traceback when stdin was closed. This
commit adds a check that stdin is not None - that indicates that stdin
is closed.