add periodic restart for ruby server
There is a memory leak in the Ruby pcsd daemon. Since we plan to remove
it entirely, there is no real benefit in looking for it and fixing it.
By swapping thin for puma server, we can restart the server on-the-fly.
This commit introduces restarting after a certain amount of requests
configurable in `/etc/sysconfig/pcsd`. The restart frees the memory, so
the memory leak is less noticeable.

The default number of requests is 200, there is a lower limit of 50 to
prevent too frequent restarts. Values that are under the lower limit
will set the lower limit. Non-integer values leave the default. It is
also possible to turn off periodic restarts by setting 0.