make url /remove_resource usable for stonith again
The internal separation of resources and stonith caused that the url
`/managec/cluster-name/remove_resource` is refusing to work with
stonith (the handler calls CLI commands that now refuses to work with
stonith). Because there is no alternative for this task, the handler is
fixed to accept stonith but only with explicit `is-stonith` parameter.