install: fix shebangs
Shebangs are generated by setuptools when generating entry point
scripts. Fedora and RHEL have py3_fix_shebangs macro in
python-rpm-macros that fixes shebangs according to packaging
guidelines. Invoking this macro guarantees the correct interpreter
path and some essential flags. The shbang macro should also be defined
on SUSE so it was left outside of the condition.

If installation is done in a virtual environment, setuptools will use
the environment's interpreter as a shebang. This is obviously incompatible
with use in RPMs, but that is fixed by the aforementioned macro. If pcs
was installed while venv was active and it is supposed to use the system
interpreter, shebangs need to be fixed manually in the installed
entrypoints. A warning is emitted during installation if venv is detected.
Detection of venv is based on PEP405.