add --no-build-isolation for dependencies
This resolves the problem of downloading new setuptools which prefer
system build dependencies over their vendored ones. Build isolation
creates a virtual environment when building a wheel where it downloads
the build requirements. However, dependencies of setuptools are not
downloaded. Depending on the age of the system, these might be
incompatible with the latest autotools. No build isolation disables the
virtual environment creation (not an issue, we use autotools to manage
dependencies) and also downloading of build requirements. This should
force pip to use everything already installed by the distribution that
is guaranteed to work together smoothly.