new build requirements
The change in build procedure requires new pip and setuptools. We will
guarantee that pcs will build with the minimum dependencies. This is
ensured via CI testing, the versions correspond to Debian 12 at this
time. You can experiment with lower versions at your own risk,
the following are the theoretical minimums:

* Setuptools 61 are needed for PEP621 support.
* Pip 19 is needed for PEP517 support.

Other changes in this patch:
* `BuildRequire:python` was removed because it is automatically
  generated by `BuildRequires:python3-devel`.
* RHEL 10+ will BuildRequire the same stuff as Fedora. RHEL 8 and 9 will
  continue with platform-python as they contain modularity which was the
  motivation for platform-python:
  https://fedoraproject.org/wiki/Changes/Platform_Python_Stack