disable build isolation for dependencies
This should help resolve mismatch in versions during CI runs. With build
isolation, when dependencies are built a new venv is created for every
dependency. This by itself is not a problem, the problem is that it also
installs build requirements that are significantly newer than the rest
of the system. This often creates errors because newer setuptools refuse
to use their vendored dependencies and the system ones are preferred.