build: fix infinite nesting in build folder
Apparently, if en exclude directive is used in tool.setuptools.packages.find
custom package discovery, this can cause infinite nesting. This was a
problem on Ubuntu, possibly caused by the presence of setuptools_scm?
This was never an issue on Fedora.
Solved thanks to https://github.com/pypa/setuptools/issues/4076.