build: fix installation of user binaries on Debian (1b/3)
This commit finishes moving the binaries which are not user-accessible.
These are copied into LIB_DIR which differs between Debian-like and
Fedora-like distros. Debian uses /usr/share while Fedora uses
/usr/lib[64]. The only difference in installation is original location
of the file, the difference in the destination path is handled by
autotools.

Therefore, there is no impact on the destination folder and the
uninstallation also relies just on the autotools LIB_DIR macro which
varies between these two distro approaches.