build: fix RECORD file based uninstall on Debian (2/3)
This part of the unistallation handles the actual pcs library. It is
located in the Python sitelib folder. There is a huge difference between
Debian and Fedora in its location. Fedora and alike uses upstream approach
with site-packages folder. Debian and alike use the same path but the
folder is called dist-packages. This is their downstream patch which
reserves dist-packages for packages that are to be used with the system
interpreter. Site-packages can be used with alternative interpreters
built from upstream.

We have to handle both paths to find the RECORD file and then prepend it
correctly to make its paths absolute. Sitelib envvar is used to store
the correct path and is also applied during deletion to make the paths
absolute.