spec: adjust scriptlets for webui backend pkg
This change in scriptlets was initiated by adding the subpackage with
webui backend. Installation, update and uninstallation of this package
needs to trigger restart of pcsd.

Additionally, the old scriptlet setup was updated to updated packaging
guidelines [1] and to respect a new systemd feature for marking services for
restart [2].

%post macros were entirely deleted as they are now automatically
interpreted by %filetriggerin macro from the systemd package. %postun
macros were traded in favor of %posttrans macros - these apply last on
upgrades and only on successful ones, so they trigger at better time
than %postun macros. The only disadvantage is when upgrading to a
package without %posttrans macro (previous versions of pcs-snmp) which
will not be restarted at all due to the now missing %postun macro.

[1]: https://docs.fedoraproject.org/en-US/packaging-guidelines/Scriptlets/
[2]: https://fedoraproject.org/wiki/Changes/Restart_services_at_end_of_rpm_transaction