fix libcurl check in upstream rpm
Turns out that the check in edc1a65 is actually substituting the pycurl
version into the Require which renders the check useless. This check is
supposed to prevent when pcsd fails to start because pycurl was built
against newer version than on the system where the rpm is installed.

I would have liked to actually substitute this from configure.ac where
pkgconfig check for libcurl is done and its version can be obtained too.
Unfortuantely, spec is not a config file and changing that for just one
substitution is not worth it.

Therefore, the dteection whether pycurl is bundled is done from
downloaded bundled sources. The libcurl version is obtained via rpm
macros which take the version from libcurl header files. This is how
it's done downstream as well.