fix url /add_meta_attr_remote for stonith
The internal separation of resources and stonith caused that the url
`/managec/cluster-name/add_meta_attr_remote` is refusing to work with
stonith (the handler calls CLI commands that now refuses to work with
stonith).
Because alternative paths (APIv2) for this task are not currently
feasible, the handler is fixed to accept stonith but only with explicit
`is-stonith` parameter.